# docs/swagger.yaml
definitions:
...
  handler.accountResponse:
    properties:
      count:
        type: integer
      id:
        type: integer
    type: object
...
  handler.contractResponse:
    properties:
      description:
        type: string
      fee:
        type: integer
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  handler.createRequest:
    properties:
      description:
        type: string
      fee:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  handler.createResponse:
    properties:
      id:
        type: integer
    type: object
...
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.getListOfContractsResponse:
    properties:
      account:
        $ref: '#/definitions/handler.accountResponse'
      contracts:
        items:
          $ref: '#/definitions/handler.contractResponse'
        type: array
    type: object
...
  handler.updateRequest:
    properties:
      description:
        type: string
      fee:
        type: integer
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: BMSTU Web Backend
  version: "1.0"
paths:
...
  /contracts:
    get:
      consumes:
      - application/json
      description: Возвращает список договоров с возможностью фильтрации по названию
        и типу договора
      parameters:
      - description: Фильтр по названию договора
        in: query
        name: contractName
        type: string
      - description: Фильтр по типу договора
        in: query
        name: contractType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список договоров и учетная запись
          schema:
            $ref: '#/definitions/handler.getListOfContractsResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение списка договоров
      tags:
      - contracts
    post:
      consumes:
      - application/json
      description: Создает новый договор с указанными данными.
      parameters:
      - description: Данные для создания договора
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание договора
          schema:
            $ref: '#/definitions/handler.createResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Создание нового договора
      tags:
      - contracts
  /contracts/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет существующий договор по его ID.
      parameters:
      - description: ID договора
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Договор успешно удален
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Договор не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Удаление договора
      tags:
      - contracts
    get:
      consumes:
      - application/json
      description: Возвращает информацию о договоре по его ID.
      parameters:
      - description: ID договора
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о договоре
          schema:
            $ref: '#/definitions/handler.contractResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Договор не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение информации о договоре
      tags:
      - contracts
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего договора по его ID.
      parameters:
      - description: ID договора
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления договора
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Договор успешно обновлен
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Договор не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Обновление данных договора
      tags:
      - contracts
  /contracts/{id}/draft:
    post:
      consumes:
      - application/json
      description: Добавляет существующий договор в текущую заявку на счёт по его
        ID.
      parameters:
      - description: ID договора
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Договор успешно добавлен в заявку на счёт
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Договор не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Добавление договора в заявку на счёт
      tags:
      - account-contracts
  /contracts/{id}/image:
    put:
      consumes:
      - multipart/form-data
      description: Обновляет изображение для договора по его ID
      parameters:
      - description: ID договора
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения для загрузки
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение успешно обновлено
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Договор не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Обновление изображения договора
      tags:
      - contracts
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"